f1_tyre_agent/requirements.txt
pydantic
langgraph
langgraph-cli[inmem]
langchain-core
langchain_openai
fastmcp
a2a-sdk
uvicorn